requirements.txt
fastapi==0.101.0
uvicorn==0.23.2
streamlit==1.31.0
sentence-transformers==2.2.2
chromadb==0.4.24
pandas==2.1.0
python-docx==0.8.11
PyPDF2==3.0.1
ollama==0.1.7
requests==2.31.0
scikit-learn==1.3.0